// packages/sdk/src/client/api.ts
export interface CompiledModel {
    /**
     * Model\'s package Name
     * @type {string}
     * @memberof CompiledModel
     */
    'packageName'?: string;
    /**
     * Model\'s relative path in its package directory.
     * @type {string}
     * @memberof CompiledModel
     */
    'path'?: string;
    /**
     * Type of malloy model file -- source file or notebook file.
     * @type {string}
     * @memberof CompiledModel
     */
    'type'?: CompiledModelTypeEnum;
    /**
     * Version of the Malloy compiler that generated the model def and results fields.
     * @type {string}
     * @memberof CompiledModel
     */
    'malloyVersion'?: string;
    /**
     * Data style for rendering query results.
     * @type {string}
     * @memberof CompiledModel
     */
    'dataStyles'?: string;
    /**
     * Malloy model def.
     * @type {string}
     * @memberof CompiledModel
     */
    'modelDef'?: string;
    /**
     * Array of model sources.
     * @type {Array<Source>}
     * @memberof CompiledModel
     */
    'sources'?: Array<Source>;
    /**
     * 
     * @type {Array<Query>}
     * @memberof CompiledModel
     */
    'queries'?: Array<Query>;
    /**
     * Array of notebook cells.
     * @type {Array<NotebookCell>}
     * @memberof CompiledModel
     */
    'notebookCells'?: Array<NotebookCell>;
}

export const CompiledModelTypeEnum = {
    Source: 'source',
    Notebook: 'notebook'
} as const;

export type CompiledModelTypeEnum = typeof CompiledModelTypeEnum[keyof typeof CompiledModelTypeEnum];
// ...
export interface NotebookCell {
    /**
     * Type of notebook cell.
     * @type {string}
     * @memberof NotebookCell
     */
    'type'?: NotebookCellTypeEnum;
    /**
     * Text contents of the notebook cell.
     * @type {string}
     * @memberof NotebookCell
     */
    'text'?: string;
    /**
     * Name of query, if this is a named query.  Otherwise, empty.
     * @type {string}
     * @memberof NotebookCell
     */
    'queryName'?: string;
    /**
     * Malloy query results. Populated only if a code cell.
     * @type {string}
     * @memberof NotebookCell
     */
    'queryResult'?: string;
}

export const NotebookCellTypeEnum = {
    Markdown: 'markdown',
    Code: 'code'
} as const;

export type NotebookCellTypeEnum = typeof NotebookCellTypeEnum[keyof typeof NotebookCellTypeEnum];
// ...
export interface Query {
    /**
     * Query\'s name.
     * @type {string}
     * @memberof Query
     */
    'name'?: string;
    /**
     * Annotations attached to query.
     * @type {Array<string>}
     * @memberof Query
     */
    'annotations'?: Array<string>;
}
// ...
export interface Source {
    /**
     * Source\'s name.
     * @type {string}
     * @memberof Source
     */
    'name'?: string;
    /**
     * Annotations attached to source.
     * @type {Array<string>}
     * @memberof Source
     */
    'annotations'?: Array<string>;
    /**
     * List of views in the source.\\
     * @type {Array<View>}
     * @memberof Source
     */
    'views'?: Array<View>;
}
/**
 * Named model view.
 * @export
 * @interface View
 */
export interface View {
    /**
     * View\'s name.
     * @type {string}
     * @memberof View
     */
    'name'?: string;
    /**
     * Annotations attached to view.
     * @type {Array<string>}
     * @memberof View
     */
    'annotations'?: Array<string>;
}